(PC-13443)[API] refactor: move activate_provider to core.providers
